assessment-tool/requirements.txt
flask
requests
psycopg2-binary
flask-limiter[redis]
google-genai
gunicorn
